+++ .eslintrc.js
@@ -13,11 +13,9 @@ module.exports = {
     'prettier',
   ],
   rules: {
-    'no-unused-vars': 'off',
     'no-console': 'warn',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     'react/no-unescaped-entities': 'off',
-
     'react/display-name': 'off',
     'react/jsx-curly-brace-presence': [
       'warn',
